Harvester refactor
* Move AWS credentals into sepertate trait
* Move HarvesterExceptions into seperate object
* Move save* methods back down to ApiHarvester until a clearer integration with FileHarvester save* methods is performed
* Add writeOut method to FileHarvester to remove duplication in Nara and P2P harvester
